% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stazioni.R
\docType{data}
\name{stazioni}
\alias{stazioni}
\title{Anagrafica delle stazioni di monitoraggio}
\format{
Un tibble co 822 righe e 27 colonne
}
\usage{
stazioni
}
\description{
Anagrafica delle stazioni di monitoraggio
}
\keyword{datasets}
